+++ release-notes.md
@@ -12,3 +12,4 @@ It should not be considered the final announcement for any release at any time.
 * Added `Promise.retry`
 * Added `Promise.delay`
 * Added `Streams.take`
+* Added `Streams.concat`